Fixes a disagreement between the behaviour and the calling code.
The calling code (see `Instruments.Probe.Runner.handle_info/2`) calls
`probe_handle_message/2`.

Any modules in the wild that rely on the runner dispatching the message
must implement the undocumented `probe_handle_message/2` instead of
`probe_handle_msg/2` and either stub that function or ignore the
warning.

Given this reasoning, the behaviour's contract was updated to specify
the function being called, this may result in new warnings for modules
that implemented the old contract but should not break any modules that
actual rely on message being dispatched correctly.

Instances in the library that use `probe_handle_msg/2` were updated to
use the correct `probe_handle_message/2`, both instances were stubs.

The tests were updated so that the test probe modules declare that they
are implementing the Instruments.Probe behaviour.