Provide better output on broken probes
Right now the probe will just crash and exit if the probe module returns
an invalid value.  With many probes this can make it difficult to find
and correct misbehaving probes.

Instead of crashing, log a warning with the probe name and the invalid
value so the probe can be corrected